put plot into container
